move interfaces around, add docs
