+++ runtime/protocol/http/common/src/aws/smithy/kotlin/runtime/http/interceptors/HttpInterceptor.kt
@@ -1,0 +1,13 @@
+/*
+ * Copyright Amazon.com, Inc. or its affiliates. All Rights Reserved.
+ * SPDX-License-Identifier: Apache-2.0
+ */
+
+package aws.smithy.kotlin.runtime.http.interceptors
+
+import aws.smithy.kotlin.runtime.client.Interceptor
+import aws.smithy.kotlin.runtime.http.request.HttpRequestBuilder
+import aws.smithy.kotlin.runtime.http.response.HttpResponse
+
+
+public typealias HttpInterceptor<I, O> = Interceptor<I, O, HttpRequestBuilder, HttpResponse>
